--- src/main/java/kuit/subway/controller/GithubApiController.java
@@ -1,88 +1,0 @@
-package kuit.subway.controller;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.springframework.web.servlet.support.RequestContextUtils;
-
-import java.io.*;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.Map;
-
-@RestController
-public class GithubApiController {
-
-    @GetMapping("/auth/github/callback")
-    public Map<Object, Object> getGithubToken(String code, RedirectAttributes redirectAttributes) throws IOException {
-        URL url = new URL("https://github.com/login/oauth/access_token");
-
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setDoInput(true);
-        conn.setDoOutput(true);
-        conn.setRequestMethod("POST");
-        conn.setRequestProperty("Accept", "application/json");
-        conn.setRequestProperty("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.45 Safari/537.36");
-
-        // 이 부분에 client_id, client_secret, code를 넣어주기
-        // client_secret 값은 주의하도록
-        try (BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(conn.getOutputStream()))) {
-            bw.write("client_id=67bb75be8f468a39c2d1&client_secret=d9485c513f8595d1dd9c499eca6aa379221f9ad5&code=" + code);
-            bw.flush();
-        }
-
-        int responseCode = conn.getResponseCode();
-
-        String responseData = getResponse(conn, responseCode);
-        System.out.println(responseData);
-
-        conn.disconnect();
-        String accessData = access(responseData, redirectAttributes);
-
-        ObjectMapper objectMapper = new ObjectMapper();
-        Map<Object, Object> result = objectMapper.readValue(accessData, Map.class);
-        return result;
-    }
-    public String access(String response, RedirectAttributes redirectAttributes) throws IOException {
-
-        ObjectMapper objectMapper = new ObjectMapper();
-        Map<String, String> map = objectMapper.readValue(response, Map.class);
-        String access_token = map.get("access_token");
-
-        URL url = new URL("https://api.github.com/user");
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setRequestMethod("GET");
-        conn.setRequestProperty("Accept", "application/json");
-        conn.setRequestProperty("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.45 Safari/537.36");
-        conn.setRequestProperty("Authorization", "token " + access_token);
-
-        int responseCode = conn.getResponseCode();
-
-        String result = getResponse(conn, responseCode);
-        System.out.println(result);
-
-        conn.disconnect();
-        redirectAttributes.addFlashAttribute("result", result);
-        return result;
-    }
-
-    private String getResponse(HttpURLConnection conn, int responseCode) throws IOException {
-        StringBuilder sb = new StringBuilder();
-        if (responseCode == 200) {
-            try (InputStream is = conn.getInputStream();
-                 BufferedReader br = new BufferedReader(new InputStreamReader(is))) {
-                for (String line = br.readLine(); line != null; line = br.readLine()) {
-                    sb.append(line);
-                }
-            }
-        }
-        return sb.toString();
-    }
-}
