+++ src/main/java/kuit/subway/dto/response/auth/TokenResponse.java
@@ -1,13 +1,17 @@
 package kuit.subway.dto.response.auth;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Getter
+@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class TokenResponse {
     private String accessToken;
+
+    public static TokenResponse of(String token){
+        return TokenResponse.builder()
+                .accessToken(token)
+                .build();
+    }
 }
