--- src/main/resources/templates/success.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-<h1>Success!</h1>
-<div th:each="key: ${result.keySet()}">
-  <span th:text="${key}"></span> = <span th:text="${result.get(key)}"></span>
-</div>
-</div>
-</body>
-</html>
